+++ pywren_ibm_cloud/job/serialize.py
@@ -14,10 +14,19 @@
 # limitations under the License.
 #
 
+import os
 import logging
+from pathlib import Path
 from io import BytesIO as StringIO
-from pywren_ibm_cloud.libs.cloudpickle import CloudPickler
-from pywren_ibm_cloud.serialize.module_dependency import ModuleDependencyAnalyzer
+from pywren_ibm_cloud.utils import bytes_to_b64str
+from pywren_ibm_cloud.libs.cloudpipe.cloudpickle import CloudPickler
+from pywren_ibm_cloud.libs.cloudpipe.module_dependency import ModuleDependencyAnalyzer
+
+try:
+    import glob2
+except Exception:
+    from pywren_ibm_cloud.libs import glob2
+
 
 logger = logging.getLogger(__name__)
 
@@ -65,3 +74,24 @@ class SerializeIndependent:
         logger.debug("Modules to transmit: {}".format(None if not mod_paths else mod_paths))
 
         return (strs, mod_paths)
+
+
+def create_module_data(mod_paths):
+
+    module_data = {}
+    # load mod paths
+    for m in mod_paths:
+        if os.path.isdir(m):
+            files = glob2.glob(os.path.join(m, "**/*.py"))
+            pkg_root = os.path.abspath(os.path.dirname(m))
+        else:
+            pkg_root = os.path.abspath(os.path.dirname(m))
+            files = [m]
+        for f in files:
+            f = os.path.abspath(f)
+            with open(f, 'rb') as file:
+                mod_str = file.read()
+            dest_filename = Path(f[len(pkg_root)+1:]).as_posix()
+            module_data[dest_filename] = bytes_to_b64str(mod_str)
+
+    return module_data
