--- pywren_ibm_cloud/libs/cloudpickle/__init__.py
@@ -1,3 +1,0 @@
-from pywren_ibm_cloud.libs.cloudpickle.cloudpickle import *
-
-__version__ = '0.6.1'
