+++ pywren_ibm_cloud/runtime/__init__.py
@@ -3,3 +3,5 @@ from pywren_ibm_cloud.runtime.deploy_utils import build_runtime
 from pywren_ibm_cloud.runtime.deploy_utils import update_runtime
 from pywren_ibm_cloud.runtime.deploy_utils import delete_runtime
 from pywren_ibm_cloud.runtime.deploy_utils import clean_runtimes
+
+from pywren_ibm_cloud.runtime.runtime import select_runtime
