--- pywren_ibm_cloud/serialize/util.py
@@ -1,54 +1,0 @@
-#
-# Copyright 2018 PyWren Team
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-"""
-Generic utility functions for serialization
-"""
-
-import os
-import base64
-from pathlib import Path
-
-try:
-    import glob2
-except:
-    from pywren_ibm_cloud.libs import glob2
-
-
-def bytes_to_b64str(byte_data):
-    byte_data_64 = base64.b64encode(byte_data)
-    byte_data_64_ascii = byte_data_64.decode('ascii')
-    return byte_data_64_ascii
-
-
-def create_mod_data(mod_paths):
-
-    module_data = {}
-    # load mod paths
-    for m in mod_paths:
-        if os.path.isdir(m):
-            files = glob2.glob(os.path.join(m, "**/*.py"))
-            pkg_root = os.path.abspath(os.path.dirname(m))
-        else:
-            pkg_root = os.path.abspath(os.path.dirname(m))
-            files = [m]
-        for f in files:
-            f = os.path.abspath(f)
-            with open(f, 'rb') as file:
-                mod_str = file.read()
-            dest_filename = Path(f[len(pkg_root)+1:]).as_posix()
-            module_data[dest_filename] = bytes_to_b64str(mod_str)
-
-    return module_data
